+++ app/src/main/java/com/themovielist/ui/recyclerview/GridStatusViewHolder.kt
@@ -1,7 +1,7 @@
 package com.themovielist.ui.recyclerview
 
-import android.support.annotation.StringRes
 import android.view.View
+import androidx.annotation.StringRes
 import com.themovielist.enums.RequestStatusDescriptor
 import kotlinx.android.synthetic.main.grid_status.view.*
 import timber.log.Timber
@@ -13,7 +13,7 @@ class GridStatusViewHolder internal constructor(itemView: View, tryAgainClick: (
         itemView.rsvRequestStatus.setTryAgainClickListener(tryAgainClick)
     }
 
-    fun bind(@RequestStatusDescriptor.RequestStatus requestStatus: Long, numberOfItems: Int) {
+    fun bind(@RequestStatusDescriptor.RequestStatus requestStatus: Int, numberOfItems: Int) {
         itemView.rsvRequestStatus.setRequestStatus(requestStatus, numberOfItems == 0)
         Timber.i("REDRAWING GRID STATUS: $requestStatus")
     }
